I think this will compile and run now.
I was making changes in the IDE but just copied the changes to the files
in the repo using an editor and things didn't change wouldn't get
detected by the editor.  This time I have made changes to the repo, made
a new project with the new codes to test it, and it will compile and run
now.  [sorry,  new to github]

changes:
-updated some new sprite coordinates
-changed spritesheet a bit
-added sound when mario bump the question block from the bottom side and
it will pop up a flower power up
-goomba and breakable rocks are now interactable

current state:
-the collision handler detected which blocks are coliding because the
power up block pop out flowers but unit movements are kinda whacked.
Maybe the current resolving collision logic is a bit off.