Fixed enemy movement [see extended notes]
Enemy units will change direction when colliding horizontally.

However, collision handler is still work in progress.  There is a
"snapping" when the two rectangles are at the corners.  Right now the
snapping happens when mario is jumpping up or falling down along the
colliding rectangle because the current collision handler checks top and
bottom collision first.  If check the sides first then mario could snap
down the ground which is not good.  However choosing the top and bottom
side collision first will result "activating" the colliding objects like
if you let mario fall next to a question block or a breakable wall while
holding the direction to that object then the object will activate what
ever it's supposed to do as if mario is bumpped under it.